Módulo 06 - Python Database/mongo_db/requirements.txt
pymongo==4.3.3
python-dotenv==1.0.0